Delete unused dao and service interfaces
