+++ service/src/main/java/com/epam/esm/certificate/CertificateService.java
@@ -1,8 +1,40 @@
 package com.epam.esm.certificate;
 
-import com.epam.esm.CrudService;
+import com.epam.esm.dao.certificate.CertificateDao;
 import com.epam.esm.model.Certificate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
-public interface CertificateService extends CrudService<Certificate> {
+import java.util.List;
+import java.util.Optional;
 
+@Service
+public class CertificateService {
+
+    private final CertificateDao certificateDao;
+
+    @Autowired
+    public CertificateService(CertificateDao certificateDao) {
+        this.certificateDao = certificateDao;
+    }
+
+    public boolean create(Certificate certificate) {
+        return certificateDao.create(certificate);
+    }
+
+    public boolean update(Certificate certificate) {
+        return certificateDao.update(certificate);
+    }
+
+    public boolean delete(long id) {
+        return certificateDao.delete(id);
+    }
+
+    public Optional<Certificate> find(long id) {
+        return certificateDao.find(id);
+    }
+
+    public List<Certificate> findAll() {
+        return certificateDao.findAll();
+    }
 }
